# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Data science competency check API.",
        "title": "Data Science Central API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Alao ramon Adebisi",
            "email": "alao.adebusy@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/MartinHeinz/go-project-blueprint/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/course/CreateCourse": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "create new course",
                "parameters": [
                    {
                        "description": "Create new course",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/modules.Course"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modules.Message"
                        }
                    }
                }
            }
        },
        "/question/CreateNewQuestion/": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "create new question",
                "parameters": [
                    {
                        "description": "create new question",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/modules.Question"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/question.RequestResponse"
                        }
                    }
                }
            }
        },
        "/question/FetchQuestionsByCourse/{StudentID}/{CourseName}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Fetch question for quis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pass email as StudentID and course name as coursename",
                        "name": "StudentID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "courseName as course name",
                        "name": "CourseName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/question.QuisRequest"
                        }
                    }
                }
            }
        },
        "/question/TestResult/": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Keeps histories of test taken by registered student",
                "parameters": [
                    {
                        "description": "Keeps histories of test taken by registered students",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/modules.TestResult"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modules.RequestResponse"
                        }
                    }
                }
            }
        },
        "/user/CreateUser": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "creates new user",
                "parameters": [
                    {
                        "description": "Create new user",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/modules.Student"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modules.ResponseManager"
                        }
                    }
                }
            }
        },
        "/user/GetUserFullInfo/{EmailAddress}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "get user full information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user emailAddress",
                        "name": "EmailAddress",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modules.Students"
                        }
                    }
                }
            }
        },
        "/user/UpdateUserDetail": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "update user details",
                "parameters": [
                    {
                        "description": "Update user information",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/modules.Student"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/modules.Message"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "modules.Course": {
            "type": "object",
            "properties": {
                "CourseCategory": {
                    "type": "string"
                },
                "CourseCode": {
                    "type": "string"
                },
                "CourseName": {
                    "type": "string"
                },
                "CourseStatus": {
                    "type": "string"
                },
                "DateCreated": {
                    "type": "string"
                },
                "PassMark": {
                    "type": "string"
                },
                "QuestionCount": {
                    "type": "string"
                }
            }
        },
        "modules.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "modules.Question": {
            "type": "object",
            "properties": {
                "CorrectOption": {
                    "type": "string"
                },
                "CourseCode": {
                    "type": "string"
                },
                "CourseName": {
                    "type": "string"
                },
                "OptionA": {
                    "type": "string"
                },
                "OptionB": {
                    "type": "string"
                },
                "OptionC": {
                    "type": "string"
                },
                "OptionD": {
                    "type": "string"
                },
                "Question": {
                    "type": "string"
                },
                "Reason": {
                    "type": "string"
                }
            }
        },
        "modules.Questions": {
            "type": "object",
            "properties": {
                "CorrectOption": {
                    "type": "string"
                },
                "CourseCode": {
                    "type": "string"
                },
                "CourseName": {
                    "type": "string"
                },
                "DateCreated": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "IsAvailable": {
                    "type": "boolean"
                },
                "OptionA": {
                    "description": "CourseCategory string `json:\"CourseCategory\" validate:\"omitempty\"`",
                    "type": "string"
                },
                "OptionB": {
                    "type": "string"
                },
                "OptionC": {
                    "type": "string"
                },
                "OptionD": {
                    "type": "string"
                },
                "Question": {
                    "type": "string"
                },
                "Reason": {
                    "type": "string"
                },
                "RecCount": {
                    "type": "integer"
                },
                "Status": {
                    "description": "Active , non-Active",
                    "type": "string"
                }
            }
        },
        "modules.RequestResponse": {
            "type": "object",
            "properties": {
                "ResponseCode": {
                    "type": "string"
                },
                "ResponseMessage": {
                    "type": "string"
                }
            }
        },
        "modules.ResponseManager": {
            "type": "object",
            "properties": {
                "ResponseCode": {
                    "type": "string"
                },
                "ResponseDescription": {
                    "type": "string"
                }
            }
        },
        "modules.Student": {
            "type": "object",
            "properties": {
                "Category": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "FirstName": {
                    "type": "string"
                },
                "Gender": {
                    "type": "string"
                },
                "LastName": {
                    "type": "string"
                },
                "MiddleName": {
                    "type": "string"
                },
                "Nationality": {
                    "type": "string"
                },
                "UserType": {
                    "type": "string"
                }
            }
        },
        "modules.Students": {
            "type": "object",
            "properties": {
                "Category": {
                    "type": "string"
                },
                "DateCreated": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "FirstName": {
                    "type": "string"
                },
                "Gender": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "LastName": {
                    "type": "string"
                },
                "MiddleName": {
                    "type": "string"
                },
                "Nationality": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "UserType": {
                    "type": "string"
                }
            }
        },
        "modules.TblCourse": {
            "type": "object",
            "properties": {
                "CourseCategory": {
                    "type": "string"
                },
                "CourseCode": {
                    "type": "string"
                },
                "CourseName": {
                    "type": "string"
                },
                "CourseStatus": {
                    "type": "string"
                },
                "DateCreated": {
                    "type": "string"
                },
                "ID": {
                    "type": "integer"
                },
                "PassMark": {
                    "type": "string"
                },
                "QuestionCount": {
                    "type": "string"
                }
            }
        },
        "modules.TestResult": {
            "type": "object",
            "properties": {
                "StudentID": {
                    "type": "string"
                },
                "TestID": {
                    "type": "string"
                },
                "TestResult": {
                    "type": "string"
                }
            }
        },
        "question.QuisRequest": {
            "type": "object",
            "properties": {
                "courseDetails": {
                    "type": "object",
                    "$ref": "#/definitions/modules.TblCourse"
                },
                "myQuestion": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/modules.Questions"
                    }
                },
                "studentDetails": {
                    "type": "object",
                    "$ref": "#/definitions/modules.Students"
                }
            }
        },
        "question.RequestResponse": {
            "type": "object",
            "properties": {
                "ResponseCode": {
                    "type": "string"
                },
                "ResponseMessage": {
                    "type": "string"
                }
            }
        }
    }
}